refactors folium_viz base system
