+++ mescal/visualizations/folium_viz_system/kpi_collection_map_visualizer.py
@@ -3,7 +3,8 @@ from typing import TYPE_CHECKING, List
 import folium
 
 from mescal.kpis import KPICollection, KPI
-from mescal.visualizations.folium_viz_system.folium_generators import FoliumObjectGenerator, TooltipGenerator
+from mescal.visualizations.folium_viz_system.folium_generators import TooltipGenerator
+from mescal.visualizations.folium_viz_system.system_base import FoliumObjectGenerator
 from mescal.visualizations.folium_viz_system.map_data_item import KPIDataItem, MapDataItem
 
 if TYPE_CHECKING:
